+++ Asteroids/Renderer.h
@@ -30,7 +30,7 @@ public:
     float getDeltaTime() { return deltaTime; }
     int getWindowHeight() { return WindowHeight; }
     int getWindowWidth() { return WindowWidth; }
-    void DrawSimpleRect(int x, int y, float width, float height, int r, int g, int b, int alpha);
+    void DrawSimpleRect(int x, int y, int width, int height, int r, int g, int b, int alpha);
     void DrawRect(Position* P1, Position* P2, Position* P3, Position* P4, int r, int g, int b, int alpha);
     void DrawTriangle(Position* P1, Position* P2, Position* P3, int r, int g, int b, int alpha);
     void DrawLine(Position* P1, Position* P2, int r, int g, int b, int alpha);
